Settings for saving tasks

+++ src/controllers/tasks.controller.js
@@ -7,8 +7,6 @@ const __dirname = path.dirname(__filename);
 
 export const controllerPOST = (req, res) => {
   const data = req.body;
-  console.log("create task", data);
-
   const folderPath = path.join(__dirname, "..", "database");
   const filePath = path.join(folderPath, "data.json");
 
@@ -23,12 +21,18 @@ export const controllerPOST = (req, res) => {
       json = JSON.parse(fileData);
     }
 
+    const taskExists = json.some((task) => task.taskName === data.taskName);
+
+    if (taskExists) {
+      return res.status(400).json({ message: "La tarea ya existe" });
+    }
+
     json.push(data);
 
     fs.writeFile(filePath, JSON.stringify(json, null, 2), (err) => {
       if (err) {
         console.error("Error al escribir el archivo:", err);
-        return res.status(500).json({ error: "Error al guardar la tarea" });
+        return res.status(500).json({ message: "Error al guardar la tarea" });
       }
       res.json({ message: "Tarea guardada con éxito" });
     });
